Refactor security management and update handling
- Updated Aspire.Hosting.PostgreSQL package version to 13.1.3.
- Added DbUpdateConcurrencyException handling in SecurityEndpoints for better error management.
- Removed unnecessary Aspire.Hosting.AppHost package reference.
- Modified DeleteSecurityAliasCommandHandler to return false instead of throwing an exception when an alias cannot be removed.
- Changed Version property type in BaseEntity from uint to long for better compatibility.
- Enhanced Security entity methods to prevent unnecessary updates and maintain timestamps when values are unchanged.
- Improved SecurityAlias entity to avoid redundant updates.
- Refactored SecurityRepository to manage entity states more effectively during updates.
- Replaced inline editing of securities with a FluentDataGrid for better UI/UX in Securities.razor.
- Introduced SecurityEditDialog component for managing security details and aliases.
- Added unit tests for Security entity methods to ensure correct behavior on updates and alias management.

+++ src/Privestio.Domain/Entities/BaseEntity.cs
@@ -26,7 +26,7 @@ public abstract class BaseEntity : IAuditableEntity
     public DateTime UpdatedAt { get; set; }
     public bool IsDeleted { get; set; }
     public DateTime? DeletedAt { get; set; }
-    public uint Version { get; set; }
+    public long Version { get; set; }
 
     public void SoftDelete()
     {
